isolate log level runtime settings
