+++ pkg/config/runtime_setting_log_level.go
@@ -1,0 +1,41 @@
+/// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2020 Datadog, Inc.
+
+package config
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/util/log"
+)
+
+// LogLevel wraps operations to change log level at runtime
+type logLevel string
+
+var ll logLevel = "log_level"
+
+func (l logLevel) Description() string {
+	return "Set/get the log level, valid values are: trace, debug, info, warn, error, critical and off"
+}
+
+func (l logLevel) Name() string {
+	return string(l)
+}
+
+func (l logLevel) Get() (interface{}, error) {
+	level, err := log.GetLogLevel()
+	if err != nil {
+		return "", err
+	}
+	return level.String(), nil
+}
+
+func (l logLevel) Set(v interface{}) error {
+	logLevel := v.(string)
+	err := log.ChangeLogLevel(logLevel)
+	if err != nil {
+		return err
+	}
+	Datadog.Set("log_level", logLevel)
+	return nil
+}
